+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Readarr support. Add instances of type `readarr` to manage book libraries alongside existing Radarr, Sonarr, and Lidarr instances.
+
 ## [0.8.0] - 2026-05-21
 
 ### Added